Add initial python file uploader
A rough but working file uploader for One Drive. Choses the correct
implementation based on the file size.

Small files can be uploaded directly:
https://docs.microsoft.com/en-us/onedrive/developer/rest-api/api/driveitem_put_content?view=odsp-graph-online

Large files need an upload session:
https://docs.microsoft.com/en-us/onedrive/developer/rest-api/api/driveitem_createuploadsession?view=odsp-graph-online

For this to work an application needs to be registered (public client
with Files.ReadWrite API permission) and the correct client_id and
tenant_id need to be used. Tenant can be passed as a commandline
argument if "common" is not the right one, client_id needs to be changed
in the source for now.